+++ gdbstub.c
@@ -1943,7 +1943,7 @@ static void gdb_chr_event(void *opaque, int event)
     }
 }
 
-static void gdb_monitor_output(GDBState *s, const char *msg, int len)
+static void gdb_monitor_output(GDBState *s, const char *msg, size_t len)
 {
     char buf[MAX_PACKET_LENGTH];
 
@@ -1954,10 +1954,10 @@ static void gdb_monitor_output(GDBState *s, const char *msg, int len)
     put_packet(s, buf);
 }
 
-static int gdb_monitor_write(Chardev *chr, const uint8_t *buf, int len)
+static ssize_t gdb_monitor_write(Chardev *chr, const uint8_t *buf, size_t len)
 {
     const char *p = (const char *)buf;
-    int max_sz;
+    size_t max_sz;
 
     max_sz = (sizeof(gdbserver_state->last_packet) - 2) / 2;
     for (;;) {
@@ -1969,7 +1969,7 @@ static int gdb_monitor_write(Chardev *chr, const uint8_t *buf, int len)
         p += max_sz;
         len -= max_sz;
     }
-    return len;
+    return (ssize_t)len;
 }
 
 #ifndef _WIN32
